move KeyManager to separate file
